backend/requirements.txt
notion-client==2.3.0
fastapi
uvicorn
matplotlib
seaborn
pandas
python-dotenv